Added a way to force result to array format
The xml2js library has no way of defining which fields should be arrays.

As a result, if an API call can return a range of options from Populi, or
only one, then the result is not deterministic.

Specifying `true` for the `forceTopLevelArray` parameter in
`taskRequest()` will determine if it is not an array and then coerce it
if necessary.